integrate eviction and pebble
