+++ core/rawdb/database.go
@@ -325,8 +325,10 @@ func NewLevelDBDatabase(file string, cache int, handles int, namespace string, r
 
 // NewPebbleDBDatabase creates a persistent key-value database without a freezer
 // moving immutable chain segments into cold storage.
+// linas
 func NewPebbleDBDatabase(file string, cache int, handles int, namespace string, readonly, ephemeral bool) (ethdb.Database, error) {
-	db, err := pebble.New(file, cache, handles, namespace, readonly, ephemeral)
+	// db, err := pebble.New(file, cache, handles, namespace, readonly, ephemeral)
+	db, err := pebble.New(10, file, file+"cold", cache, handles, namespace, readonly, ephemeral)
 	if err != nil {
 		return nil, err
 	}
