+++ ethdb/eviction/lru/lru.go
@@ -2,11 +2,13 @@ package lru
 
 import (
 	"container/list"
+	"sync"
 )
 
 type Eviction struct {
 	lruList list.List
 	elemmap map[string]*list.Element
+	mu      sync.Mutex
 }
 
 // Constructor
@@ -19,6 +21,8 @@ func New() *Eviction {
 // Return evition victim (least recently used one) and true
 // Return nil and false if there is no element in policy
 func (e *Eviction) SelectVictim() ([]byte, bool) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 	elem := e.lruList.Front()
 	if elem == nil {
 		return nil, false
@@ -30,6 +34,8 @@ func (e *Eviction) SelectVictim() ([]byte, bool) {
 // Update access information of key and return true if key exists
 // Otherwise, return false
 func (e *Eviction) Access(key []byte) bool {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 	strKey := string(key)
 	elem := e.elemmap[strKey]
 	if elem == nil {
@@ -43,6 +49,8 @@ func (e *Eviction) Access(key []byte) bool {
 // Return eviction victim  and true, and remove it from policy
 // Return nil and false if there is no element in policy
 func (e *Eviction) Pop() ([]byte, bool) {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 	elem := e.lruList.Front()
 	if elem == nil {
 		return nil, false
@@ -54,9 +62,18 @@ func (e *Eviction) Pop() ([]byte, bool) {
 	return []byte(strKey), true
 }
 
+// Number of keys in policy
+func (e *Eviction) Len() int {
+	e.mu.Lock()
+	defer e.mu.Unlock()
+	return len(e.elemmap)
+}
+
 // Add new key to policy and return true
 // return false if key already exists
 func (e *Eviction) Push(key []byte) bool {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 	strKey := string(key)
 	val := e.elemmap[strKey]
 	if val != nil {
@@ -71,6 +88,8 @@ func (e *Eviction) Push(key []byte) bool {
 // Remove key from policy and return true
 // Return false if key does not exist
 func (e *Eviction) Delete(key []byte) bool {
+	e.mu.Lock()
+	defer e.mu.Unlock()
 	strKey := string(key)
 	elem := e.elemmap[strKey]
 	if elem == nil {
