+++ ethdb/pebble_original/pebble.go
@@ -15,7 +15,7 @@
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
 // Package pebble implements the key-value database layer based on pebble.
-package pebble_modified
+package pebble_original
 
 import (
 	"bytes"
@@ -24,9 +24,6 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
-	"syscall"
-	// "unsafe"
-	// "runtime/debug"
 
 	"github.com/cockroachdb/pebble"
 	"github.com/cockroachdb/pebble/bloom"
@@ -54,18 +51,12 @@ const (
 	degradationWarnInterval = time.Minute
 )
 
-var overThresholdFlag = false
-const path = "."
-
 // Database is a persistent key-value store based on the pebble storage engine.
 // Apart from basic data storage functionality it also supports batch writes and
 // iterating over the keyspace in binary-alphabetical order.
 type Database struct {
-	hotFn string     // filename for reporting
-	coldFn string     // filename for reporting
-	hotDb *pebble.DB // Underlying pebble storage engine
-	coldDb *pebble.DB
-	ssdThreshold int
+	fn string     // filename for reporting
+	db *pebble.DB // Underlying pebble storage engine
 
 	compTimeMeter       metrics.Meter // Meter for measuring the total time spent in database compaction
 	compReadMeter       metrics.Meter // Meter for measuring the data read during compaction
@@ -102,22 +93,6 @@ type Database struct {
 
 	writeOptions *pebble.WriteOptions
 }
-
-func getDiskUsage(path string) (total uint64, free uint64, used uint64, usage float64, err error) {
-	fs := syscall.Statfs_t{}
-	err = syscall.Statfs(path, &fs)
-	if err != nil {
-		return
-	}
-
-	total = fs.Blocks * uint64(fs.Bsize)
-	free = fs.Bfree * uint64(fs.Bsize)
-	used = total - free
-	usage = (float64(used) / float64(total)) * 100
-
-	return
-}
-
 
 func (d *Database) onCompactionBegin(info pebble.CompactionInfo) {
 	if d.activeComp == 0 {
@@ -169,7 +144,7 @@ func (l panicLogger) Fatalf(format string, args ...interface{}) {
 
 // New returns a wrapped pebble DB object. The namespace is the prefix that the
 // metrics reporting should use for surfacing internal stats.
-func New(ssdThreshold int, file1, file2 string, cache int, handles int, namespace string, readonly bool, ephemeral bool) (*Database, error) {
+func New(file string, cache int, handles int, namespace string, readonly bool, ephemeral bool) (*Database, error) {
 	// Ensure we have some minimal caching and file guarantees
 	if cache < minCache {
 		cache = minCache
@@ -177,7 +152,7 @@ func New(ssdThreshold int, file1, file2 string, cache int, handles int, namespac
 	if handles < minHandles {
 		handles = minHandles
 	}
-	logger := log.New("database", file1)
+	logger := log.New("database", file)
 	logger.Info("Allocated cache and file handles", "cache", common.StorageSize(cache*1024*1024), "handles", handles)
 
 	// The max memtable size is limited by the uint32 offsets stored in
@@ -207,9 +182,7 @@ func New(ssdThreshold int, file1, file2 string, cache int, handles int, namespac
 		memTableSize = maxMemTableSize - 1
 	}
 	db := &Database{
-		hotFn:           file1,
-		coldFn:           file2,
-		ssdThreshold:   ssdThreshold,
+		fn:           file,
 		log:          logger,
 		quitChan:     make(chan chan error),
 		writeOptions: &pebble.WriteOptions{Sync: !ephemeral},
@@ -261,16 +234,11 @@ func New(ssdThreshold int, file1, file2 string, cache int, handles int, namespac
 	opt.Experimental.ReadSamplingMultiplier = -1
 
 	// Open the db and recover any potential corruptions
-	innerDB1, err := pebble.Open(file1, opt)
+	innerDB, err := pebble.Open(file, opt)
 	if err != nil {
 		return nil, err
 	}
-	db.hotDb = innerDB1
-	innerDB2, err := pebble.Open(file2, opt)
-	if err != nil {
-		return nil, err
-	}
-	db.coldDb = innerDB2
+	db.db = innerDB
 
 	db.compTimeMeter = metrics.GetOrRegisterMeter(namespace+"compact/time", nil)
 	db.compReadMeter = metrics.GetOrRegisterMeter(namespace+"compact/input", nil)
@@ -309,38 +277,19 @@ func (d *Database) Close() error {
 		}
 		d.quitChan = nil
 	}
-	err1 := d.hotDb.Close()
-	err2 := d.coldDb.Close()
-	if err1 != nil {
-		return err1
-	}
-	if err2 != nil {
-		return err2
-	}
-	return nil
+	return d.db.Close()
 }
 
 // Has retrieves if a key is present in the key-value store.
 func (d *Database) Has(key []byte) (bool, error) {
-	// fmt.Printf("h:k: %X\n", key)
 	d.quitLock.RLock()
 	defer d.quitLock.RUnlock()
 	if d.closed {
 		return false, pebble.ErrClosed
 	}
-	fmt.Println("Has in hot db")
-	_, closer, err := d.hotDb.Get(key)
+	_, closer, err := d.db.Get(key)
 	if err == pebble.ErrNotFound {
-		// check cold db if key is not found in hot db
-		fmt.Println("Has in cold db")
-		_, closer, err := d.coldDb.Get(key)
-		if err == pebble.ErrNotFound {
-			return false, nil
-		}else if err != nil {
-			return false, err
-		}
-		defer closer.Close()
-		return true, nil
+		return false, nil
 	} else if err != nil {
 		return false, err
 	}
@@ -350,134 +299,61 @@ func (d *Database) Has(key []byte) (bool, error) {
 
 // Get retrieves the given key if it's present in the key-value store.
 func (d *Database) Get(key []byte) ([]byte, error) {
-	// fmt.Printf("g:k: %X\n", key)
 	d.quitLock.RLock()
 	defer d.quitLock.RUnlock()
 	if d.closed {
 		return nil, pebble.ErrClosed
 	}
-	fmt.Println("Get in hot db")
-	dat, closer, err := d.hotDb.Get(key)
+	dat, closer, err := d.db.Get(key)
 	if err != nil {
-		// check cold db if key is not found in hot db
-		fmt.Println("Get in cold db")
-		dat, closer, err := d.coldDb.Get(key)
-		if err != nil {
-			return nil, err
-		}
-		ret := make([]byte, len(dat))
-		copy(ret, dat)
-		closer.Close()
-		return ret, nil
+		return nil, err
 	}
 	ret := make([]byte, len(dat))
 	copy(ret, dat)
 	closer.Close()
 	return ret, nil
 }
+
 // Put inserts the given value into the key-value store.
 func (d *Database) Put(key []byte, value []byte) error {
-	// fmt.Printf("p:k: %X v: %d\n", key, int(unsafe.Sizeof(value)))
-	// debug.PrintStack()
 	d.quitLock.RLock()
 	defer d.quitLock.RUnlock()
 	if d.closed {
 		return pebble.ErrClosed
 	}
-	
-	total, free, used, usage, err := getDiskUsage(path)
-	if err != nil {
-		fmt.Println("Error:", err)
-		return err
-	}
-	fmt.Printf("Total: %d bytes\nFree: %d bytes\nUsed: %d bytes\nUsage: %.2f%%\n", total, free, used, usage)
-	
-	if usage >= float64(d.ssdThreshold) {
-		overThresholdFlag = true
-	} else {
-		overThresholdFlag = false
-	}
-
-	if overThresholdFlag {
-		// lru.eviction
-		fmt.Println("Eviction")
-	}
-	return d.hotDb.Set(key, value, d.writeOptions)
-}
-
-// Put inserts the given value into the key-value store.
-func (d *Database) PutForTest(key []byte, value []byte) error {
-	d.quitLock.RLock()
-	defer d.quitLock.RUnlock()
-	if d.closed {
-		return pebble.ErrClosed
-	}
-	// total, free, used, usage, err := getDiskUsage(path)
-	_, _, _, usage, err := getDiskUsage(path)
-	if err != nil {
-		fmt.Println("Error:", err)
-		return err
-	}
-	fmt.Printf("Usage: %.2f%%\nThreshold: %d%%\n",usage,d.ssdThreshold)
-	
-	if usage >= float64(d.ssdThreshold) {
-		overThresholdFlag = true
-	} else {
-		overThresholdFlag = false
-	}
-
-	if overThresholdFlag {
-		fmt.Println("Put in cold db")
-		return d.coldDb.Set(key, value, d.writeOptions)
-	}else{
-		fmt.Println("Put in hot db")
-		return d.hotDb.Set(key, value, d.writeOptions)
-	}
+	return d.db.Set(key, value, d.writeOptions)
 }
 
 // Delete removes the key from the key-value store.
 func (d *Database) Delete(key []byte) error {
-	// fmt.Printf("d:k: %X\n", key)
 	d.quitLock.RLock()
 	defer d.quitLock.RUnlock()
 	if d.closed {
 		return pebble.ErrClosed
 	}
-	if err := d.hotDb.Delete(key, d.writeOptions); err != nil {
-		return err
-	}
-	if err := d.coldDb.Delete(key, d.writeOptions); err != nil {
-		return err
-	}
-	return nil
+	return d.db.Delete(key, nil)
 }
-
 
 // NewBatch creates a write-only key-value store that buffers changes to its host
 // database until a final write is called.
 func (d *Database) NewBatch() ethdb.Batch {
 	return &batch{
-		bHot:  d.hotDb.NewBatch(),
-		bCold: d.coldDb.NewBatch(),
+		b:  d.db.NewBatch(),
 		db: d,
 	}
 }
-
 
 // NewBatchWithSize creates a write-only database batch with pre-allocated buffer.
 func (d *Database) NewBatchWithSize(size int) ethdb.Batch {
 	return &batch{
-		bHot:  d.hotDb.NewBatchWithSize(size),
-		bCold:  d.coldDb.NewBatchWithSize(size),
+		b:  d.db.NewBatchWithSize(size),
 		db: d,
 	}
 }
 
-
 // snapshot wraps a pebble snapshot for implementing the Snapshot interface.
 type snapshot struct {
-	dbHot *pebble.Snapshot
-	dbCold *pebble.Snapshot
+	db *pebble.Snapshot
 }
 
 // NewSnapshot creates a database snapshot based on the current state.
@@ -486,69 +362,42 @@ type snapshot struct {
 // Note don't forget to release the snapshot once it's used up, otherwise
 // the stale data will never be cleaned up by the underlying compactor.
 func (d *Database) NewSnapshot() (ethdb.Snapshot, error) {
-	snapHot := d.hotDb.NewSnapshot()
-	snapCold := d.coldDb.NewSnapshot()
-	return &snapshot{dbHot: snapHot, dbCold: snapCold}, nil
+	snap := d.db.NewSnapshot()
+	return &snapshot{db: snap}, nil
 }
-// Has checks if the given key is present in either the hot or cold snapshot.
+
+// Has retrieves if a key is present in the snapshot backing by a key-value
+// data store.
 func (snap *snapshot) Has(key []byte) (bool, error) {
-	// Check in hot snapshot
-	if has, err := snap.hasInSnapshot(snap.dbHot, key); err != nil || has {
-		fmt.Println("Has in snapshot of hot db")
-		return has, err
-	}
-
-	// Check in cold snapshot
-	fmt.Println("Has in snapshot of cold db")
-	return snap.hasInSnapshot(snap.dbCold, key)
-}
-
-// hasInSnapshot is a helper function to check if a key exists in a given snapshot.
-func (snap *snapshot) hasInSnapshot(snapDb *pebble.Snapshot, key []byte) (bool, error) {
-	_, closer, err := snapDb.Get(key)
+	_, closer, err := snap.db.Get(key)
 	if err != nil {
 		if err != pebble.ErrNotFound {
 			return false, err
+		} else {
+			return false, nil
 		}
-		return false, nil
 	}
 	closer.Close()
 	return true, nil
 }
 
-// Get retrieves the given key if it's present in either the hot or cold snapshot.
+// Get retrieves the given key if it's present in the snapshot backing by
+// key-value data store.
 func (snap *snapshot) Get(key []byte) ([]byte, error) {
-	// Try to get from hot snapshot
-	if data, err := snap.getFromSnapshot(snap.dbHot, key); err == nil {
-		fmt.Println("Get in snapshot of hot db")
-		return data, nil
-	} else if err != pebble.ErrNotFound {
-		return nil, err
-	}
-
-	// Try to get from cold snapshot
-	fmt.Println("Get in snapshot of cold db")
-	return snap.getFromSnapshot(snap.dbCold, key)
-}
-
-// getFromSnapshot is a helper function to get a key from a given snapshot.
-func (snap *snapshot) getFromSnapshot(snapDb *pebble.Snapshot, key []byte) ([]byte, error) {
-	dat, closer, err := snapDb.Get(key)
+	dat, closer, err := snap.db.Get(key)
 	if err != nil {
 		return nil, err
 	}
-	defer closer.Close()
-
 	ret := make([]byte, len(dat))
 	copy(ret, dat)
+	closer.Close()
 	return ret, nil
 }
 
 // Release releases associated resources. Release should always succeed and can
 // be called multiple times without causing error.
 func (snap *snapshot) Release() {
-	snap.dbHot.Close()
-	snap.dbCold.Close()
+	snap.db.Close()
 }
 
 // upperBound returns the upper bound for the given prefix
@@ -567,11 +416,9 @@ func upperBound(prefix []byte) (limit []byte) {
 }
 
 // Stat returns the internal metrics of Pebble in a text format. It's a developer
-// method to read everything there is to read independent of Pebble version.
-//
-// The property is unused in Pebble as there's only one thing to retrieve.
-func (d *Database) Stat(property string) (string, error) {
-	return d.hotDb.Metrics().String(), nil
+// method to read everything there is to read, independent of Pebble version.
+func (d *Database) Stat() (string, error) {
+	return d.db.Metrics().String(), nil
 }
 
 // Compact flattens the underlying data store for the given key range. In essence,
@@ -593,22 +440,12 @@ func (d *Database) Compact(start []byte, limit []byte) error {
 	if limit == nil {
 		limit = bytes.Repeat([]byte{0xff}, 32)
 	}
-	// TODO: Parallelization Compaction?
-	// Compact both hot and cold databases
-	if err := d.hotDb.Compact(start, limit, true); err != nil {
-		return err
-	}
-
-	if err := d.coldDb.Compact(start, limit, true); err != nil {
-		return err
-	}
-
-	return nil
+	return d.db.Compact(start, limit, true) // Parallelization is preferred
 }
 
 // Path returns the path to the database directory.
 func (d *Database) Path() string {
-	return d.hotFn
+	return d.fn
 }
 
 // meter periodically retrieves internal pebble counters and reports them to
@@ -638,7 +475,7 @@ func (d *Database) meter(refresh time.Duration, namespace string) {
 			compRead  int64
 			nWrite    int64
 
-			stats              = d.hotDb.Metrics()
+			stats              = d.db.Metrics()
 			compTime           = d.compTime.Load()
 			writeDelayCount    = d.writeDelayCount.Load()
 			writeDelayTime     = d.writeDelayTime.Load()
@@ -724,27 +561,21 @@ func (d *Database) meter(refresh time.Duration, namespace string) {
 // batch is a write-only batch that commits changes to its host database
 // when Write is called. A batch cannot be used concurrently.
 type batch struct {
-	bHot    *pebble.Batch
-	bCold    *pebble.Batch
+	b    *pebble.Batch
 	db   *Database
 	size int
 }
 
 // Put inserts the given value into the batch for later committing.
 func (b *batch) Put(key, value []byte) error {
-	//linas
-	// fmt.Printf("bp:k: %X v: %d\n", key, int(unsafe.Sizeof(value)))
-	// debug.PrintStack()
-	b.bHot.Set(key, value, nil)
+	b.b.Set(key, value, nil)
 	b.size += len(key) + len(value)
 	return nil
 }
 
 // Delete inserts the key removal into the batch for later committing.
 func (b *batch) Delete(key []byte) error {
-	// fmt.Printf("bd:k: %X\n", key)
-	b.bHot.Delete(key, nil)
-	b.bCold.Delete(key, nil)
+	b.b.Delete(key, nil)
 	b.size += len(key)
 	return nil
 }
@@ -761,32 +592,18 @@ func (b *batch) Write() error {
 	if b.db.closed {
 		return pebble.ErrClosed
 	}
-	if err := b.bHot.Commit(b.db.writeOptions); err != nil {
-		fmt.Println("dbHot Batch Write Error:", err)
-		return err
-	}
-	return nil
+	return b.b.Commit(b.db.writeOptions)
 }
 
 // Reset resets the batch for reuse.
 func (b *batch) Reset() {
-	b.bHot.Reset()
-	b.bCold.Reset()
+	b.b.Reset()
 	b.size = 0
 }
 
+// Replay replays the batch contents.
 func (b *batch) Replay(w ethdb.KeyValueWriter) error {
-	// Replay hot batch
-	if err := replayBatch(b.bHot, w); err != nil {
-		fmt.Println("dbHot Batch Replay Error:", err)
-		return err
-	}
-	return nil
-}
-
-// replayBatch replays the operations in the given batch to the KeyValueWriter.
-func replayBatch(batch *pebble.Batch, w ethdb.KeyValueWriter) error {
-	reader := batch.Reader()
+	reader := b.b.Reader()
 	for {
 		kind, k, v, ok, err := reader.Next()
 		if !ok || err != nil {
@@ -810,158 +627,58 @@ func replayBatch(batch *pebble.Batch, w ethdb.KeyValueWriter) error {
 //
 // The pebble iterator is not thread-safe.
 type pebbleIterator struct {
-	iterHot     *pebble.Iterator
-	iterCold    *pebble.Iterator
-	movedHot    bool
-	movedCold   bool
-	releasedHot bool
-	releasedCold bool
-	validHot    bool
-	validCold   bool
-	turnHot     bool
+	iter     *pebble.Iterator
+	moved    bool
+	released bool
 }
 
 // NewIterator creates a binary-alphabetical iterator over a subset
 // of database content with a particular key prefix, starting at a particular
 // initial key (or after, if it does not exist).
 func (d *Database) NewIterator(prefix []byte, start []byte) ethdb.Iterator {
-	iterHot, _ := d.hotDb.NewIter(&pebble.IterOptions{
+	iter, _ := d.db.NewIter(&pebble.IterOptions{
 		LowerBound: append(prefix, start...),
 		UpperBound: upperBound(prefix),
 	})
-	iterCold, _ := d.coldDb.NewIter(&pebble.IterOptions{
-		LowerBound: append(prefix, start...),
-		UpperBound: upperBound(prefix),
-	})
-	iterHot.First()
-	iterCold.First()
-	return &pebbleIterator{
-		iterHot:     iterHot,
-		iterCold:    iterCold,
-		movedHot:    true,
-		movedCold:   true,
-		releasedHot: false,
-		releasedCold: false,
-		validCold:  true,
-		validHot:   true,
-		turnHot:    true, // current turn
-	}
+	iter.First()
+	return &pebbleIterator{iter: iter, moved: true, released: false}
 }
 
+// Next moves the iterator to the next key/value pair. It returns whether the
+// iterator is exhausted.
 func (iter *pebbleIterator) Next() bool {
-	if iter.movedHot && iter.movedCold {
-		iter.movedHot = false
-		iter.movedCold = false
-		iter.validHot = iter.iterHot.Valid()
-		iter.validCold = iter.iterCold.Valid()		
-		if iter.validHot && iter.validCold {
-			result := bytes.Compare(iter.KeyHot(), iter.KeyCold())
-			if result > 0 {
-				iter.turnHot = false
-			}else if result < 0 {
-				iter.turnHot = true
-			}else {
-				iter.turnHot = true
-				iter.validCold = iter.iterCold.Next()
-			}
-		}else if iter.validHot {
-			iter.turnHot = true
-		}else if iter.validCold {
-			iter.turnHot = false
-		}
-		return iter.validHot || iter.validCold
+	if iter.moved {
+		iter.moved = false
+		return iter.iter.Valid()
 	}
-
-	if iter.turnHot {
-		iter.validHot = iter.iterHot.Next()
-	} else {
-		iter.validCold = iter.iterCold.Next()
-	}
-	
-	if iter.validHot && iter.validCold {
-		result := bytes.Compare(iter.KeyHot(), iter.KeyCold())
-		if result > 0 {
-			iter.turnHot = false
-		}else if result < 0 {
-			iter.turnHot = true
-		}else {
-			if iter.turnHot {
-				iter.validCold = iter.iterCold.Next()
-			}else {
-				iter.validHot = iter.iterHot.Next()
-			}
-			if iter.validHot {
-				iter.turnHot = true
-			}else {
-				iter.turnHot = false
-			}
-		}
-	}else if iter.validHot {
-		iter.turnHot = true
-	}else if iter.validCold {
-		iter.turnHot = false
-	}
-	
-	return iter.validHot || iter.validCold
+	return iter.iter.Next()
 }
 
 // Error returns any accumulated error. Exhausting all the key/value pairs
 // is not considered to be an error.
 func (iter *pebbleIterator) Error() error {
-	if err := iter.iterHot.Error(); err != nil {
-		return err
-	}
-	return iter.iterCold.Error()
+	return iter.iter.Error()
 }
 
 // Key returns the key of the current key/value pair, or nil if done. The caller
 // should not modify the contents of the returned slice, and its contents may
 // change on the next call to Next.
 func (iter *pebbleIterator) Key() []byte {
-	if iter.turnHot {
-		return iter.iterHot.Key()
-	} else {
-		return iter.iterCold.Key()
-	}
-	return nil
-}
-
-func (iter *pebbleIterator) KeyHot() []byte {
-	if iter.validHot {
-		return iter.iterHot.Key()
-	}
-	return nil
-}
-
-func (iter *pebbleIterator) KeyCold() []byte {
-	if iter.validCold {
-		return iter.iterCold.Key()
-	}
-	return nil
+	return iter.iter.Key()
 }
 
 // Value returns the value of the current key/value pair, or nil if done. The
 // caller should not modify the contents of the returned slice, and its contents
 // may change on the next call to Next.
 func (iter *pebbleIterator) Value() []byte {
-	if iter.turnHot {
-		return iter.iterHot.Value()
-	} else {
-		return iter.iterCold.Value()
-	}
-	return nil
+	return iter.iter.Value()
 }
 
 // Release releases associated resources. Release should always succeed and can
 // be called multiple times without causing error.
 func (iter *pebbleIterator) Release() {
-	if !iter.releasedHot {
-		iter.iterHot.Close()
-		iter.releasedHot = true
-	}
-	if !iter.releasedCold {
-		iter.iterCold.Close()
-		iter.releasedCold = true
+	if !iter.released {
+		iter.iter.Close()
+		iter.released = true
 	}
 }
-
